KUBIG CONFERENCE/Recommend system/Lastfm-Spotify-recommender-system-using-bandit/requirements.txt
streamlit
requests
python-dotenv
spotipy